raycasting v1

+++ srcs/draw_pixels/my_pixel_put.c
@@ -1,10 +1,10 @@
 #include "cub3d.h"
 
-void	my_pixel_put(t_img_data *data, int x, int y, int color)
+void	my_pixel_put(t_img_data *data, float x, float y, int color)
 {
 	char	*dst;
 
-	dst = data->addr + (y * data->line_length + x * (data->bits_per_pixel / 8));
+	dst = data->addr + ((int)y * data->line_length + (int)x * (data->bits_per_pixel / 8));
 	*(unsigned int *)dst = color;
 }
